moved sui-build dependent parts into sui-move-build
