+++ crates/sui-cluster-test/src/test_case/fullnode_build_publish_transaction_test.rs
@@ -4,8 +4,8 @@
 use crate::{TestCaseImpl, TestContext};
 use async_trait::async_trait;
 use jsonrpsee::rpc_params;
-use sui_core::test_utils::compile_basics_package;
 use sui_json_rpc_types::SuiTransactionBlockEffectsAPI;
+use sui_move_build::test_utils::compile_basics_package;
 use sui_types::{base_types::ObjectID, object::Owner};
 
 pub struct FullNodeBuildPublishTransactionTest;
