+++ crates/sui-move-build/src/lib.rs
@@ -58,6 +58,27 @@ use sui_verifier::verifier as sui_bytecode_verifier;
 #[path = "unit_tests/build_tests.rs"]
 mod build_tests;
 
+pub mod test_utils {
+    use crate::{BuildConfig, CompiledPackage, SuiPackageHooks};
+    use std::path::PathBuf;
+
+    pub fn compile_basics_package() -> CompiledPackage {
+        compile_example_package("../../examples/move/basics")
+    }
+
+    pub fn compile_managed_coin_package() -> CompiledPackage {
+        compile_example_package("../../crates/sui-core/src/unit_tests/data/managed_coin")
+    }
+
+    pub fn compile_example_package(relative_path: &str) -> CompiledPackage {
+        move_package::package_hooks::register_package_hooks(Box::new(SuiPackageHooks));
+        let mut path = PathBuf::from(env!("CARGO_MANIFEST_DIR"));
+        path.push(relative_path);
+
+        BuildConfig::new_for_testing().build(&path).unwrap()
+    }
+}
+
 /// Wrapper around the core Move `CompiledPackage` with some Sui-specific traits and info
 #[derive(Debug, Clone)]
 pub struct CompiledPackage {
